Add support for forcing aarch64 builds.
This will allow doing some initial tests on slackwareaarch64-current. If
there is a binary repackage, it will currently expect the downloads to
be in DOWNLOAD_x86_64.

This might be a little counter-intuitive, but there are enough binary
repackaging builds in the repo that do not support x86 such that if we
suddenly start adding DOWNLOAD links to them, we have to remove
UNSUPPORTED from the field, which will cause more problems than it is
worth.